pyproject.toml
[tool.poetry]
name = "paperblog"
version = "0.1.0"
description = "digital version of my notebook"
authors = ["navanchauhan <navanchauhan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
configparser = "^5.3.0"
Pillow = "^9.3.0"
PyMuPDF = "^1.20.2"
tqdm = "^4.64.1"
azure-cognitiveservices-vision-computervision = "^0.9.0"
Jinja2 = "^3.1.2"
markdown2 = "^2.4.6"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
